Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,10 +1,9 @@
 // Your code here...
 #include <stdio.h>
-
 int main(){
-    int a,b;
+    float a,b;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%f %f %c",&a,&b,&c);
     if(c=='+'){
         printf("%d",a+b);}
     else if(c=='-'){
@@ -12,7 +11,7 @@ int main(){
     else if(c=='*'){
         printf("%d",a*b);}
     else{
-        printf("%d",a/b);
+        printf("%f",a/b);
     } 
     return 0;
 }
